move event related classes to more fitting packages
Forecast in Material Disposition
https://github.com/metasfresh/metasfresh/issues/2643
