+++ de.metas.material/dispo-service/src/main/java/de/metas/material/dispo/service/event/DistributionPlanEventHandler.java
@@ -17,10 +17,10 @@ import de.metas.material.dispo.DemandCandidateDetail;
 import de.metas.material.dispo.DistributionCandidateDetail;
 import de.metas.material.dispo.service.CandidateChangeHandler;
 import de.metas.material.dispo.service.event.SupplyProposalEvaluator.SupplyProposal;
-import de.metas.material.event.DistributionPlanEvent;
 import de.metas.material.event.EventDescr;
 import de.metas.material.event.ddorder.DDOrder;
 import de.metas.material.event.ddorder.DDOrderLine;
+import de.metas.material.event.ddorder.DistributionPlanEvent;
 import lombok.NonNull;
 
 /*
