+++ de.metas.material/dispo-service/src/test/java/de/metas/material/dispo/service/event/ForecastEventHandlerTest.java
@@ -27,11 +27,11 @@ import de.metas.material.dispo.model.X_MD_Candidate;
 import de.metas.material.dispo.service.CandidateChangeHandler;
 import de.metas.material.dispo.service.StockCandidateFactory;
 import de.metas.material.event.EventDescr;
-import de.metas.material.event.ForecastEvent;
 import de.metas.material.event.MaterialDemandEvent;
 import de.metas.material.event.MaterialDescriptor;
 import de.metas.material.event.MaterialEventService;
 import de.metas.material.event.forecast.Forecast;
+import de.metas.material.event.forecast.ForecastEvent;
 import de.metas.material.event.forecast.ForecastLine;
 import lombok.NonNull;
 import mockit.Delegate;
