+++ de.metas.material/event/src/main/java/de/metas/material/event/pporder/ProductionPlanEvent.java
@@ -1,9 +1,10 @@
-package de.metas.material.event;
+package de.metas.material.event.pporder;
 
 import org.adempiere.util.lang.impl.TableRecordReference;
-import org.eevolution.model.I_DD_Order;
+import org.eevolution.model.I_PP_Order;
 
-import de.metas.material.event.ddorder.DDOrder;
+import de.metas.material.event.EventDescr;
+import de.metas.material.event.MaterialEvent;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -31,16 +32,17 @@ import lombok.NonNull;
  * #L%
  */
 /**
- * Send by the material planner when it came up with a distribution plan that could be turned into an {@link I_DD_Order}.
+ * Send by the material planner when it came up with a brilliant production plan that could be turned into an {@link I_PP_Order} <b>or</or> if a ppOrder was actually created or changed.
  *
  * @author metas-dev <dev@metasfresh.com>
+ *
  */
 @Data
-@Builder
 @AllArgsConstructor
-public class DistributionPlanEvent implements MaterialEvent
+@Builder
+final public class ProductionPlanEvent implements MaterialEvent
 {
-	public static final String TYPE = "DistributionPlanEvent";
+	public static final String TYPE = "ProductionPlanEvent";
 
 	@NonNull
 	private final EventDescr eventDescr;
@@ -48,18 +50,5 @@ public class DistributionPlanEvent implements MaterialEvent
 	private final TableRecordReference reference;
 
 	@NonNull
-	private final DDOrder ddOrder;
-
-
-	/**
-	 * Note: this field is a bit redundant because the {@link #getPpOrder()}'s lines contain a network distribution line with this info. However, the material-dispo code doesn't know or care about how to get to that information.
-	 */
-	@NonNull
-	private final Integer fromWarehouseId;
-
-	/**
-	 * Also check the note about {@link #getFromWarehouseId()}.
-	 */
-	@NonNull
-	private final Integer toWarehouseId;
+	private final PPOrder ppOrder;
 }
