+++ de.metas.material/planning/src/main/java/de/metas/material/planning/pporder/PPOrderPojoConverter.java
@@ -3,9 +3,9 @@ package de.metas.material.planning.pporder;
 import org.eevolution.model.I_PP_Order_BOMLine;
 import org.springframework.stereotype.Service;
 
-import de.metas.material.event.PPOrderRequestedEvent;
 import de.metas.material.event.pporder.PPOrder;
 import de.metas.material.event.pporder.PPOrderLine;
+import de.metas.material.event.pporder.PPOrderRequestedEvent;
 import lombok.NonNull;
 
 /*
